Fix errrors in network. (linux and windows C++ things arent the same [I HATE BOTH OF YOU WHY NOT STANDARDIZE!]) revert "&" checks on type instead of "=="

+++ dll/steam_matchmaking_servers.cpp
@@ -641,7 +641,7 @@ void Steam_Matchmaking_Servers::server_details(Gameserver *g, gameserveritem_t *
                 g->set_num_players(ssq_a2s_info->players);
                 g->set_version(static_cast<uint32_t>(std::stoull(ssq_a2s_info->version, NULL, 0)));
                 if (ssq_info_has_port(ssq_a2s_info)) g->set_port(ssq_a2s_info->port);
-                if (ssq_info_has_gameid(ssq_a2s_info)) g->set_appid(static_cast<uint32_t>(ssq_a2s_info->gameid));
+                if (ssq_info_has_gameid(ssq_a2s_info)) g->set_appid(ssq_a2s_info->gameid);
                 else g->set_appid(ssq_a2s_info->id);
                 g->set_offline(false);
             } else {
